+++ citelang/main/client.py
@@ -102,11 +102,13 @@ class Client(base.BaseClient):
         root = self._graph(*args, **kwargs)
         return results.Graph(root).graph(fmt=fmt)
 
-    def badge(self, *args, **kwargs):
+    def badge(self, template="sunburst", *args, **kwargs):
         """
         Generate a badge for a package
         """
         root = self._graph(*args, **kwargs)
+        if template == "treemap":
+            return results.Treemap(root)
         return results.Badge(root)
 
     def _graph(