+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on GitHub.
 
 ## [0.0.x](https://github.com/vsoch/citelang/tree/main) (0.0.x)
+ - expanding interactive badge into treemap and sunburst (0.0.14)
  - adding basic citelang gen to create markdown files for repos (and actions) (0.0.13)
  - parsing markdown to generate credit table (0.0.12)
  - support for citelang badge to generate interactive html page (0.0.11)